Rename constants file to plans-constants
