+++ packages/calypso-products/src/index.ts
@@ -1,7 +1,7 @@
 export * from './main';
 export * from './types';
 export * from './plans-utilities';
-export * from './constants';
+export * from './plans-constants';
 export * from './jetpack-constants';
 export * from './utils';
 export * from './product-values';
