+++ packages/calypso-products/src/is-titan-mail.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { formatProduct } from './format-product';
-import { TITAN_MAIL_MONTHLY_SLUG } from './constants';
+import { TITAN_MAIL_MONTHLY_SLUG } from './plans-constants';
 
 export function isTitanMail( product ) {
 	product = formatProduct( product );
