+++ packages/calypso-products/test/get-interval-type-for-term.js
@@ -3,7 +3,7 @@
  */
 
 import { getIntervalTypeForTerm } from '../src/get-interval-type-for-term';
-import { TERM_ANNUALLY, TERM_BIENNIALLY, TERM_MONTHLY } from '../src/constants';
+import { TERM_ANNUALLY, TERM_BIENNIALLY, TERM_MONTHLY } from '../src/plans-constants';
 
 describe( 'getIntervalTypeForTerm', () => {
 	test( 'should return 2-year intervalType if current plan is a 2-year plan', () => {
