+++ packages/calypso-products/test/get-popular-plan-spec.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { getPopularPlanSpec } from '../src';
-import { GROUP_WPCOM, TYPE_BUSINESS, TYPE_PREMIUM } from '../src/constants';
+import { GROUP_WPCOM, TYPE_BUSINESS, TYPE_PREMIUM } from '../src/plans-constants';
 
 describe( 'getPopularPlanSpec()', () => {
 	const availablePlans = [
